Added exception handling to data and web layers

+++ src/data/explorer.py
@@ -1,6 +1,6 @@
-from .init import curs, conn
+from .init import curs, conn, IntegrityError
 from model.explorer import Explorer
-
+from .errors import MissingException, DuplicateException
 
 curs.execute("""create table if not exists explorer(
                 name text primary key,
@@ -14,27 +14,44 @@ def model_to_dict(explorer: Explorer) -> dict | None:
     return explorer.model_dump() if explorer else None
 
 def get_one(name: str) -> Explorer:
+    if not name: return None
+
     qry = "select * from explorer where name=:name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+
+    row = row_to_model(curs.fetchone())
+    if not row:
+        raise MissingException(
+            f"Explorer {name} not found"
+        )
+
+    return row
 
 def get_all() -> list[Explorer]:
     qry = "select * from explorer"
     curs.execute(qry)
     return [row_to_model(row) for row in curs.fetchall()]
 
-def create(explorer: Explorer) -> Explorer:
+def create(explorer: Explorer) -> Explorer | None:
+    if not explorer: return None
     qry = """insert into explorer (name, country, description)
              values (:name, :country, :description)"""
     params = model_to_dict(explorer)
-    _ = curs.execute(qry, params)
 
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise DuplicateException(
+            f"Explorer {explorer.name} already exists"
+        )
+    
     conn.commit()
 
     return get_one(explorer.name)
 
-def modify(explorer: Explorer) -> Explorer:
+def modify(explorer: Explorer) -> Explorer | None:
+    if not explorer: return None
     qry = """update explorer
              set country=:country,
              name=:name,
@@ -42,15 +59,25 @@ def modify(explorer: Explorer) -> Explorer:
              where name=:name_orig"""
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    conn.commit()
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    curs.execute(qry, params)
+    
+    if curs.rowcount == 1:
+        conn.commit()
+        return get_one(explorer.name)
+    else:
+        raise MissingException(msg=f"Explorer {explorer.name} not found")
 
 def delete(name: str) -> bool:
+    if not name: return False
+
     qry = "delete from explorer where name = :name"
     params = {"name": name}
-    res = curs.execute(qry, params)
+    curs.execute(qry, params)
+
+    if curs.rowcount != 1:
+        raise MissingException(
+            f"Explorer {name} not found"
+        )
     conn.commit()
     
-    return bool(res)
+    return True
